Added Files Via Upload
Had to do a lot of updates, didn't realize this code was no longer being pushed to GitHub, 

Removed style.css, it was just messing with white-theme and bootstrap. 

Bugs:
When clicking on the seizure logs on index.html it will display a "deletion success!" message when no files are deleted. It seems to be listening and pushing the message for any 'click' in the table. 

Dark Mode still doesn't stick to cache.